pyproject.toml
[tool.poetry]
name = "youpost-backend"
version = "0.1.0"
description = ""
authors = ["Tim <timvash90@gmail.com>"]
readme = "README.md"
packages = [{include = "application"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.17"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "application.main:start"